Improve attachment.py

+++ plugins/module_utils/attachment.py
@@ -7,6 +7,7 @@ from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
 
+import hashlib
 import mimetypes
 import os
 
@@ -18,8 +19,7 @@ def _path(*subpaths):
 
 
 def _query(original=None):
-    # Flatten the response (skip embedded links to resources)
-    return dict(original or {}, sysparm_exclude_reference_link="true")
+    return dict(original or {})
 
 
 class AttachmentClient:
@@ -49,7 +49,7 @@ class AttachmentClient:
 
         return result
 
-    def get_record(self, query, sys_id=None, file=True, must_exist=False):
+    def get_record(self, query, sys_id=None, file=False, must_exist=False):
         records = self.list_records(query, sys_id, file)
 
         if len(records) > 1:
@@ -66,8 +66,8 @@ class AttachmentClient:
 
     def create_record(self, payload, data, check_mode, mime_type=None):
         if check_mode:
-            # Approximate the result using the payload and data. TODO
-            return dict(payload, data=data)
+            # Approximate the result using the payload and data.
+            return payload
         return self.client.request_binary(
             "POST", _path("file"), data, _query(payload), mime_type
         ).json["result"]
@@ -77,24 +77,17 @@ class AttachmentClient:
         #
         # "file_dict" is a dict with a mandatory key "path" and optional "name", "type" and "encryption_context".
         # These properties can be read directly from the yaml attachment list and can be set by the user.
-        if "name" in file_dict and file_dict["name"] is not None:
-            payload["file_name"] = file_dict["name"]
-        else:
-            payload["file_name"] = os.path.splitext(
-                os.path.basename(file_dict["path"])
-            )[0]
+        payload["file_name"] = get_file_name(file_dict)
+        file_type = get_file_type(file_dict)
 
-        if "type" in file_dict and file_dict["type"] is not None:
-            file_type = file_dict["type"]
-        else:
-            file_type = mimetypes.guess_type(file_dict["path"])[0]
-
-        if "encryption_context" in file_dict and file_dict["encryption_context"] is not None:
+        if (
+            "encryption_context" in file_dict
+            and file_dict["encryption_context"] is not None
+        ):
             payload["encryption_context"] = file_dict["encryption_context"]
 
-        file_obj = open(file_dict["path"], "rb")
-        data = file_obj.read()
-        file_obj.close()
+        with open(file_dict["path"], "rb") as file_obj:
+            data = file_obj.read()
         return self.create_record(payload, data, check_mode, file_type)
 
     def upload_records(self, payload, file_dict_list, check_mode):
@@ -103,6 +96,71 @@ class AttachmentClient:
             for file_dict in (file_dict_list or [])
         ]
 
-    def delete_record(self, record, check_mode):
+    def update_record(self, payload, file_dict, check_mode):
+        if self.is_changed(payload, file_dict):
+            self.delete_record(
+                self.get_record(build_query(payload, file_dict)), check_mode, True
+            )
+            return dict(
+                {"changed": True, "msg": "Changes detected, hash doesn't match remote."},
+                **self.upload_record(payload, file_dict, check_mode)
+            )
+        else:
+            return dict(
+                {"changed": False, "msg": "Skipped. Hash matches remote."},
+                **self.get_record(build_query(payload, file_dict))
+            )
+
+    def update_records(self, payload, file_dict_list, check_mode):
+        return [
+            self.update_record(payload, file_dict, check_mode)
+            for file_dict in (file_dict_list or [])
+        ]
+
+    def delete_record(self, record, check_mode, silent=False):
         if not check_mode:
-            self.client.delete(_path(record["sys_id"]))
+            if record is None and silent:
+                return {"changed": False, "msg": "Skipped. Record doesn't exist."}
+            else:
+                self.client.delete(_path(record["sys_id"]))
+
+    def is_changed(self, payload, file_dict):
+        rec = self.get_record(build_query(payload, file_dict))
+        if rec is not None:
+            return file_hash(file_dict["path"]) != rec["hash"]
+        return True
+
+    def are_changed(self, payload, file_dict_list):
+        if file_dict_list is not None:
+            return [
+                self.is_changed(payload, file_dict)
+                for file_dict in (file_dict_list or [])
+            ]
+        return []
+
+
+def get_file_name(file_dict):
+    if "name" in file_dict and file_dict["name"] is not None:
+        return file_dict["name"]
+    else:
+        return os.path.splitext(os.path.basename(file_dict["path"]))[0]
+
+
+def get_file_type(file_dict):
+    if "type" in file_dict and file_dict["type"] is not None:
+        return file_dict["type"]
+    else:
+        return mimetypes.guess_type(file_dict["path"])[0]
+
+
+def build_query(payload, file_dict):
+    return dict(
+        file_name=get_file_name(file_dict),
+        content_type=get_file_type(file_dict),
+        **payload
+    )
+
+
+def file_hash(file):
+    with open(file, "rb") as f:
+        return hashlib.sha256(f.read()).hexdigest()
